Move shapelib file into ogr/ogrsf_frmts/shape/shapelib/
