+++ ogr/ogrsf_frmts/avc/CMakeLists.txt
@@ -25,10 +25,13 @@ add_gdal_driver(
 gdal_standard_includes(ogr_AVC)
 
 if (NOT OGR_ENABLE_DRIVER_SHAPE OR OGR_ENABLE_DRIVER_SHAPE_PLUGIN)
-target_compile_definitions(ogr_AVC PRIVATE -DWITHOUT_SHAPEFILE)
+    target_compile_definitions(ogr_AVC PRIVATE -DWITHOUT_SHAPEFILE)
+    target_include_directories(ogr_AVC PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../shape)
+    target_include_directories(ogr_AVC PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../shape/shapelib)
 else()
     if (GDAL_USE_SHAPELIB_INTERNAL)
       gdal_add_vendored_lib(ogr_AVC shapelib)
+      target_include_directories(ogr_AVC PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../shape/shapelib)
     else ()
       gdal_target_link_libraries(ogr_AVC PRIVATE SHAPELIB::shp)
     endif ()
