+++ ogr/ogrsf_frmts/avc/avc_bin.cpp
@@ -172,9 +172,9 @@
 #define VSI_SHP_WriteMoreDataOK OGRAVC_VSI_SHP_WriteMoreDataOK
 #define SASetupDefaultHooks OGRAVC_SASetupDefaultHooks
 
-#include "../shape/shapefil.h"
-#include "../shape/dbfopen.c"
-#include "../shape/shp_vsi.c"
+#include "shapefil.h"
+#include "dbfopen.c"
+#include "shp_vsi.c"
 
 #else
