+++ ogr/ogrsf_frmts/shape/shapelib/README.TXT
@@ -1,0 +1,1 @@
+Upstream for those .c and .h files is https://github.com/OSGeo/shapelib
