+++ ogr/ogrsf_frmts/shape/shapelib/dbfopen.c
