+++ ogr/ogrsf_frmts/shape/shapelib/sbnsearch.c
