+++ ogr/ogrsf_frmts/shape/shapelib/shapefil_private.h
