+++ ogr/ogrsf_frmts/shape/shapelib/shpopen.c
