+++ ogr/ogrsf_frmts/shape/shapelib/shptree.c
